chore: move 2023 to base route
