+++ src/pages/sustain-africa.js
@@ -1,12 +1,12 @@
 import React from "react";
 
-import SEO from "../../components/SEO/seo";
-import Nav from "../../components/2023/nav";
-import Layout from "../../components/layout";
+import SEO from "../components/SEO/seo";
+import Nav from "../components/2023/nav";
+import Layout from "../components/layout";
 
 // import { homeBackground } from "../../components/2023/cloudImages";
-import Footer from "../../components/2023/footer";
-import Breadcrumb from "../../components/2023/breadcrumb";
+import Footer from "../components/2023/footer";
+import Breadcrumb from "../components/2023/breadcrumb";
 
 const SectionLayout = ({ children }) => (
   <div className="mb-[56px] px-[24px] lg:px-0 font-helvetica  max-w-[795px] m-auto">{children}</div>
